src/analysis/predictive/response.py
"""
Predictive analysis for community responses.
"""

from typing import Dict
import pandas as pd
from sklearn.ensemble import RandomForestRegressor
from ..base import BaseAnalyzer

class ResponsePredictor(BaseAnalyzer):
    """Predicts community response patterns"""
    
    def train_engagement_model(self, training_data: pd.DataFrame):
        """Train the engagement prediction model"""
        pass

    def predict_post_engagement(self, post_data: Dict) -> Dict:
        """Predict engagement for a new post"""
        pass
