requirements.txt
praw==7.7.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
vaderSentiment==3.3.2
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0
networkx==3.2.1
scikit-learn==1.3.0
